bug fix
signature change was wrong in sample app : pane was in with of 1 px
merge LINUX and APPLE preprocessor defines with UNIX

+++ ImGuiFileDialog/ImGuiFileDialog.h
@@ -92,7 +92,7 @@ private:
 	size_t dlg_optionsPaneWidth = 0;
 	std::string searchTag;
 	UserDatas dlg_userDatas;
-	int dlg_CountSelectionMax = 1; // 0 for infinite
+	size_t dlg_CountSelectionMax = 1; // 0 for infinite
 
 public:
 	static ImGuiFileDialog* Instance()
